+++ src/commands/usage.go
@@ -23,6 +23,7 @@ func PrintUsage() {
 	fmt.Printf("  %s             Verify system health and check for tool requirements\n", green("check"))
 	fmt.Printf("  %s                (Still in development) Manage language versions (e.g. Go, Python) via subcommands\n", green("vm"))
 	fmt.Printf("  %s           Install system-wide packages or tools (e.g. 'zoi install node@18')\n", green("install"))
+	fmt.Printf("  %s         Uninstall system-wide packages or tools (e.g. 'zoi uninstall node')\n", green("uninstall"))
 	fmt.Printf("  %s            Create a new application from a predefined template\n", green("create"))
 	fmt.Printf("  %s              Generate an application from a local YAML configuration file\n", green("make"))
 	fmt.Printf("  %s               Execute a command defined in a local zoi.yaml file\n", green("run"))